implement RemoteField

+++ src/RemoteField.h
@@ -5,7 +5,7 @@
 class RemoteField : public Field
 {
 public:
-    RemoteField(sf::Vector2i pos, const sf::Font& fnt);
+    RemoteField(sf::Vector2f pos, const sf::Font& fnt);
 
     void setHit(sf::Vector2i pos, int hit);
 };
